+++ Changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project follows to [Ragnarök Versioning Convention](https://shor.cz/ragnarok_versioning_convention).
 
-## [Unreleased] Wither Config Version 1 Changelog
+## Wither Config Version 1 Changelog - 2023-09-06
 
 ### Added
 
@@ -14,6 +14,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - The Wither follow range
   - The Wither armor
   - The Wither follow distance
+  - The Wither unarmored fly height
   - The Wither summoning sequence length
   - The Wither summoning sequence end explosion strength
+  - The Wither skulls damage
+  - The Wither skulls magic damage
+  - The Wither skulls heal towards the Wither when getting a kill
   - The Wither skulls explosion strength
+  - The Wither skulls potion effects
+  - Whether the Wither should attempt to break blocks around it when targeting a player
+- Fix for the Vanilla broken Wither target following behavior